+++ src/CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ### {{version_pkg_lupo}} ({{creationDate_changelog}})
+  - add responsive columns to [lupo] shortcode view
+
+### 4.11.1 (February 27, 2026)
   - plg_editors-xtd_lupotoy improvements
   - add missing translations for plg_editors-xtd_lupotoy